check generated patch via diff
We want to ensure that the 001 patch committed in the repository for
each branch is the most up to date patch so that when it is used to
alter the Go tree during RPM builds it will always apply cleanly and
produce the correct code.

We enforce this by issuing a `git diff` command with the `--exit-code`
flag to fail the CI step if a diff is present.