Remove command lowering Go version before go mod tidy
We did this historically because when using the host Go toolchain to run
commands such as `go mod tidy` within the `go/src` directory it would
fail because the host toolchain was lower than the new source toolchain.
We now build the source toolchain first and then use it directly, effectively
ignoring the host toolchain so this is no longer a problem.